requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-compositekey==1.5.0
django-filter==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-jsonapi==2.7.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
djoser==1.7.0
gunicorn==19.9.0
inflection==0.3.1
Markdown==3.1.1
psycopg2==2.8.2
PyJWT==1.7.1
pytz==2019.1
Serializer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
